Squashed commit of fixmass branch.
Switching to a new default normalization whereby both mass and
radius are held fixed for each density profile, by renormalizing
density after HE shape is determined. The 1-bar density becomes
an observable included in the loss function. Since the estimated
delta on rho0 is about 1% (i.e. large) this makes mixing easier.
An additional benefit is easier comparison with other groups and
removal of ambigiuity about which mass normalizes the Js and MoI.

For these reasons, fixed mass is the new default and --M-error
will be meaningless. The chi^2 degrees of freedom remain the same.
The old behavior can be reproduced with --fixed-mass=0 but this
will probably be deprecated and removed in the future. Existing
samples might be replaced too. Future samples should be assumed
to be of fixed mass unless indicated otherwise.