Adding period loss function for period-non-preseriving sampling.
Note, like mass and rho0 the losses.period is *always* added to
the loss dsqr. It will be near zero if args.preserve_period is
True (default) and otherwise should be included in the chi2 df
count. Note that this is different from directly sampling rotation
with mrot being a MCMC parameter. That option still exists but
will probably be removed soon. It is probably not a good idea to
mix these methods of allowing rotation uncertainty. In other words
it's probably best to stick to --fix-rot being True. The dof count
is confusing because when mrot was sampled directly it was part of
the prior while now we treat rotation as part of the likelihood.