Prepping for sampling with varying rotation.
I think I changed my mind and am going to keep using a sampled m
parameter with a normal prior and preserving rotation period on the
implied rotation period by obs.s0. Rotation is not part of the
likelihood and does not count towards dof (but should included in
the subtracted prior if subtracted from output L). I make the
sample-space m a standard normal and use obs to transform to proper
mu and sig before sending to generic_priors.rotation. This is all
it takes. Oh and putting a zero in front of the quadratic base
seeds. Now I'll restart the sampling from the quadratic seed.