Remove lingering traces of mrot as input MCMC parameter.
I still get surprised by how bad it is to use mrot as if we know
anything about it a priori. Even after all this time. Basically,
the value in obs.s0 is nothing more than a crude guess.

Changed generic_priors.rotation to use obs.P, and removed hopefully
all reliance on m in _lnprob. If --fix-rot=0 then the first number
in x is a standard normal to be mapped to the rotation prior. The
resulting Prot is put in obs.P and the TOFPlanet is relaxed to that
rotation period. The rotation period appears also in the loss
function but should contribute zero if --preserve-period (default
and soon only option). The rotation period should cound as a chi2
dof when looking at the likelihood - prior, so do not remove it
from the prior.