Set Prot in a tp instead of modifying an obs.
In driver and planet cooker, removed the risky practice of setting
obs.P = Prot before calling TOFPlanet(obs). Since obs is mutable
this has easy to miss side effects when making more than one
planet. It is unclear if we need to re-run previous samples though,
perhaps emcee kept a local copy of obs in each walker.

TOFPlanet modified to only calculate wrot and mrot when needed,
not in set_observables. Only period is an observable. And mrot and
qrot removed from ascii exports to prevent possible ambiguity; let
the importer calculate it!