Implemented custom r(J2,J4,J6) loss function of French.
Implemented as a dynamically eval'ed call to an undocumented
losses method. The eval must use a simple string instead of a
pre-compiled code object because it must be pickled and sent to
worker cores. But pre-compiling seems to have neglegible effect
on speed anyway; it's probably in the noise.
Now to get a sample with this likelihood.