Squashed commit of branch bennotropes.
This commit implements barotrope-based parameterization sampling.
The core functions are in module ppbs.py and the assiciated driver
is drive-ppbs.py. The relax_to_barotrope method is newly
implemented in TOFPlanet.py and synced with the TOF-planet repo.

The ppbs module is similar to ppwd but simplified and streamlined.
Currently the ppbs parameterization is always a triple-polytrope
with parameters [K1,n1,K2,n2,K3,n3,r12,r23], with fixed rotation
period taken from observables. Alternatively rotation period can
be a sampled variable, tacked at the *end* of the parameter vector.
(Note ppwd puts the rotation period in the beginning, which is
awkward but necessary since the polytnomial degree can vary.)

The model builder is ppwd_planet(N,x,obs,toforder,xlevels). As the
name suggests, it returns the TOFPlanet object directly, with the
critical fields populated. The returned tp is ready for the call
tp.relax_to_barotrope(). The starting density profile is quadratic
but it makes no difference.

The transform, untransform, and other helper functions are
available as quasi-private methods. A uniform prior on generous
supports is implemented; I expect more informative priors should be
implemented in the future.

The driver is very similar to the ppwd driver and the interface is
the same. Some obsolete options are left undocumented and will be
removed later. Some defaults are modified to better fit the
barotrope relaxation. Using fixed mass and preserved period (both
default True) is highly recommended.

Max density implemented as loss instead of prior since it is not
available pre-relaxation.

Planet cooker and planets_from_sample script implemented and
renamed to distinguish ppwd from ppbs.

Matlab equivalents are partially implemented. Everything needed to
run samples is avaiable; additional analysis and plotting tools
are coming soon.